nest MissionsPanel components
