+++ src/web/components/MissionsPanel/MissionSetStorage/DeleteMissionSetButton/DeleteMissionSetButton.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
-import { missionSet } from "../../../data/mission_set/mission-set";
-import { DialogActions } from "../../../types/context-types";
-import { listSavedMissionSets } from "../../../utils/local-storage";
+import { missionSet } from "../../../../data/mission_set/mission-set";
+import { DialogActions } from "../../../../types/context-types";
+import { listSavedMissionSets } from "../../../../utils/local-storage";
 import { DisabledCodes } from "./delete-messages";
 import { DeleteMissionSetDialog } from "./DeleteMissionSetDialog";
