+++ src/web/components/MissionsPanel/MissionSetStorage/LoadMissionSetButton/LoadMissionSetButton.tsx
@@ -1,8 +1,8 @@
 import { useContext, useState } from "react";
-import { JaiaDispatchContext } from "../../../context/JaiaContext";
-import { JaiaActions } from "../../../context/jaia-actions";
-import { DialogActions } from "../../../types/context-types";
-import { listSavedMissionSets } from "../../../utils/local-storage";
+import { JaiaDispatchContext } from "../../../../context/JaiaContext";
+import { JaiaActions } from "../../../../context/jaia-actions";
+import { DialogActions } from "../../../../types/context-types";
+import { listSavedMissionSets } from "../../../../utils/local-storage";
 import { DisabledCodes } from "./load-messages";
 import { LoadMissionSetDialog } from "./LoadMissionSetDialog";
