+++ src/web/components/MissionsPanel/MissionSetStorage/MissionSetStorage.less
@@ -1,4 +1,4 @@
-@import "../../style/stylesheets/vars.less";
+@import "../../../style/stylesheets/vars.less";
 
 .jaia-dialog.mission-set-storage {
     align-items: center;
