+++ src/web/components/MissionsPanel/MissionSpeedSliders/MissionSpeedSliders.less
